requirements.txt
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-friendship==1.8.0
django-heroku==0.3.1
django-registration-redux==1.3
gunicorn==19.9.0
Pillow==5.2.0
psycopg2==2.7.5
python-decouple==3.1
pytz==2018.5
whitenoise==4.1
